Add investment and divestment functionality (#13)
### Link to issue number:
None.

### Summary of the pull request:
This PR adds the ability to move funds between bank and investments. It also adds names to the URLconf for linking in templates. @asriniv2 you probably want to see this diff as it shows how to add similar functionality.

### Testing performed:
Tested by manually visiting the /invest and /divest URLs as well as through the links on the index page. Verified that funds moved as requested.

### Known issues with pull request:
The forms have no styling and minimal text, someone sighted may want to fix this in a follow-up PR. Now that we have Bootstrap, perhaps you can use some of their CSS classes?