Add ability to pay down credit card balances. (#14)
### Link to issue number:
None.

### Summary of the pull request:
This PR adds the ability to pay off credit card balances. It also adds a tip to always display when the user has a past-due balance. During payment, overdue (interest-accruing) balance is prioritized over the portion of the balance that does not yet accrue interest.

The payment link is hidden unless the user has a balance. Additionally, this PR hides the divest link if the user has no investments.

### Testing performed:
Tested by manipulating the values in the admin and using the view to pay off the balance.

### Known issues with pull request:
None.